Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/securitypolicyviolation/script-sample-no-opt-in-expected.txt

CONSOLE MESSAGE: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 80: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 70: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 75: Refused to execute a script because 'unsafe-eval' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 1: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT Inline script should not have a sample. Test timed out
TIMEOUT Inline event handlers should not have a sample. Test timed out
TIMEOUT JavaScript URLs in iframes should not have a sample. Test timed out
TIMEOUT eval()-alikes should not have a sample. Test timed out
